Dynamically discover end of memory

+++ src/arch.rs
@@ -1,6 +1,6 @@
 use core::arch::asm;
 
-pub fn hart_id() -> u64 {
+pub fn hart_id() -> usize {
     let id;
     unsafe {
         asm!("mv {0}, tp", out(reg) id);
